Move the join channel form to its own component
